feat(summarize): implement structured outputs and modular refactoring
- Upgrade @anthropic-ai/sdk from v0.36.3 to v0.71.2
- Implement Structured Outputs API for guaranteed valid JSON (fixes parsing errors)
- Extract summarize.ts into 6 focused modules (223 lines, down from 1,742)
  - batchProcessing.ts: stateful batch processing logic (447 lines)
  - speakerContributions.ts: speaker contribution generation (316 lines)
  - compression.ts: ID compression/decompression (121 lines)
  - prompts.ts: all prompt templates (429 lines)
  - types.ts: type definitions (73 lines)
  - utils.ts: utility functions (151 lines)
- Create reusable usage logging utilities (logUsage, logMultiPhaseUsage)
- Track Phase 3 enrichment LLM usage (was missing)
- Add administrativeBodyName parameter support for generic organ types
- Update prompts to support δημοτικές επιτροπές and δημοτικές κοινότητες
- Replace 3 hardcoded NO_USAGE instances with imports
- Update all usage tracking to use new utilities
