+++ package.json
@@ -16,7 +16,7 @@
     "smoke": "tsc && node dist/cli.js smoke"
   },
   "dependencies": {
-    "@anthropic-ai/sdk": "^0.36.3",
+    "@anthropic-ai/sdk": "^0.71.2",
     "@schemalabs/diavgeia-cli": "^0.1.0",
     "aws-sdk": "^2.1687.0",
     "axios": "^1.7.7",
