+++ src/lib/subjectEnrichment.ts
@@ -1,6 +1,8 @@
 import { Subject, SubjectContext } from '../types.js';
 import { geocodeLocation } from './geocode.js';
 import { getSubjectContextWithClaude } from './claudeSearch.js';
+import { ResultWithUsage, NO_USAGE } from './ai.js';
+import Anthropic from '@anthropic-ai/sdk';
 
 /**
  * Input shape for subject enrichment, compatible with both ExtractedSubject and SubjectInProgress
@@ -33,13 +35,13 @@ export interface EnrichmentConfig {
  * @param input - Subject data to enrich
  * @param id - Subject ID (generated by caller based on their strategy)
  * @param config - Configuration including city name, body name, and date
- * @returns Complete Subject object with enriched data
+ * @returns Complete Subject object with enriched data and LLM usage stats
  */
 export async function enrichSubjectData(
     input: EnrichmentInput,
     id: string,
     config: EnrichmentConfig
-): Promise<Subject> {
+): Promise<ResultWithUsage<Subject>> {
     // Step 1: Geocode location if provided
     let location: Subject['location'] = null;
     if (input.locationText) {
@@ -60,14 +62,17 @@ export async function enrichSubjectData(
 
     // Step 2: Fetch web context using Claude API with web search
     let context: SubjectContext;
+    let usage: Anthropic.Messages.Usage;
     try {
-        context = await getSubjectContextWithClaude({
+        const contextResult = await getSubjectContextWithClaude({
             subjectName: input.name,
             subjectDescription: input.description,
             cityName: config.cityName,
             administrativeBodyName: config.administrativeBodyName || "Δημοτικό Συμβούλιο",
             date: config.date
         });
+        context = contextResult.result;
+        usage = contextResult.usage;
     } catch (error) {
         console.error("Error fetching subject context:", error);
         // Return empty context on error
@@ -75,20 +80,24 @@ export async function enrichSubjectData(
             text: "",
             citationUrls: []
         };
+        usage = NO_USAGE;
     }
 
     // Step 3: Construct complete Subject object
     return {
-        id,
-        name: input.name,
-        description: input.description,
-        agendaItemIndex: input.agendaItemIndex,
-        introducedByPersonId: input.introducedByPersonId,
-        speakerContributions: input.speakerContributions,
-        topicImportance: input.topicImportance,
-        proximityImportance: input.proximityImportance,
-        location,
-        topicLabel: input.topicLabel,
-        context
+        result: {
+            id,
+            name: input.name,
+            description: input.description,
+            agendaItemIndex: input.agendaItemIndex,
+            introducedByPersonId: input.introducedByPersonId,
+            speakerContributions: input.speakerContributions,
+            topicImportance: input.topicImportance,
+            proximityImportance: input.proximityImportance,
+            location,
+            topicLabel: input.topicLabel,
+            context
+        },
+        usage
     };
 }
